+++ features/step_definitions/action_link_steps.rb
@@ -3,5 +3,5 @@ Then /^I should see a member link to "([^"]*)"$/ do |name|
 end
 
 Then /^I should not see a member link to "([^"]*)"$/ do |name|
-  page.should_not have_css("a.member_link", :text => name)
+  %{Then I should not see "#{name}" within "a.member_link"}
 end