+++ features/step_definitions/configuration_steps.rb
@@ -1,10 +1,11 @@
 module ActiveAdminReloading
 
   def load_active_admin_configuration(configuration_content)
+    ActiveAdmin::Event.dispatch ActiveAdmin::Application::BeforeLoadEvent, ActiveAdmin.application
     eval(configuration_content)
-    ActiveAdmin::Event.dispatch ActiveAdmin::Application::LoadEvent, ActiveAdmin.application
+    ActiveAdmin::Event.dispatch ActiveAdmin::Application::AfterLoadEvent, ActiveAdmin.application
     Rails.application.reload_routes!
-    ActiveAdmin.application.namespaces.values.each{|n| n.load_menu! }
+    ActiveAdmin.application.namespaces.values.each{|n| n.reset_menu! }
   end
 
 end
@@ -46,7 +47,6 @@ end
 
 Given /^a configuration of:$/ do |configuration_content|
   load_active_admin_configuration(configuration_content)
-  ActiveAdmin.application.namespaces.values.each{|n| n.load_menu! }
 end
 
 Given /^an index configuration of:$/ do |configuration_content|