+++ features/step_definitions/factory_steps.rb
@@ -13,6 +13,14 @@ Given /^a (published )?post with the title "([^"]*)" written by "([^"]*)" exists
   Post.create! :title => title, :author => author, :published_at => published_at
 end
 
+Given /^(\d+) posts? written by "([^"]*)" exist$/ do |count, author_name|
+  first, last = author_name.split(' ')
+  author = User.find_or_create_by_first_name_and_last_name(first, last, :username => author_name.gsub(' ', '').underscore)
+  (0...count.to_i).each do |i|
+    Post.create! :title => "Hello World #{i}", :author => author
+  end
+end
+
 Given /^(\d+)( published)? posts? exists?$/ do |count, published|
   (0...count.to_i).each do |i|
     Post.create! :title => "Hello World #{i}", :published_at => (published ? Time.now : nil)
@@ -34,4 +42,8 @@ Given /^I create a new post with the title "([^"]*)"$/ do |title|
   click_link "New Post"
   fill_in :title, :with => title
   click_button "Create Post"
+end
+
+Given /^a store named "([^"]*)" exists$/ do |name|
+  Store.create! :name => name
 end