+++ lib/active_admin/dependency_checker.rb
@@ -25,7 +25,7 @@ module ActiveAdmin
 
       def sass_rails_3_1?
         require 'sass/rails/version'
-        Sass::Rails::VERSION >= "3.1"
+        ::Sass::Rails::VERSION >= "3.1"
       rescue LoadError
         false
       end