+++ lib/active_admin/resource_controller/collection.rb
@@ -49,7 +49,7 @@ module ActiveAdmin
             table_column = (column =~ /\./) ? column :
               "#{table}.#{active_admin_config.resource_quoted_column_name(column)}"
 
-            chain.order("#{table_column} #{order}")
+            chain.reorder("#{table_column} #{order}")
           else
             chain # just return the chain
           end
@@ -91,6 +91,9 @@ module ActiveAdmin
           if current_scope
             @before_scope_collection = chain
             scope_chain(current_scope, chain)
+          elsif active_admin_config.scope_to
+            @before_scope_collection = scoped_collection
+            chain
           else
             chain
           end