+++ spec/unit/resource_spec.rb
@@ -62,48 +62,6 @@ module ActiveAdmin
       end
     end
 
-    describe "menu item name" do
-      it "should be the resource name when not set" do
-        config.menu_item_name.should == "Categories"
-      end
-      it "should be settable" do
-        config.menu :label => "My Label"
-        config.menu_item_name.should == "My Label"
-      end
-    end
-
-    describe "parent menu item name" do
-      it "should be nil when not set" do
-        config.parent_menu_item_name.should == nil
-      end
-      it "should return the name if set" do
-        config.tap do |c|
-          c.menu :parent => "Blog"
-        end.parent_menu_item_name.should == "Blog"
-      end
-    end
-    
-    describe "menu item priority" do
-      it "should be 10 when not set" do
-        config.menu_item_priority.should == 10
-      end
-      it "should be settable" do
-        config.menu :priority => 2
-        config.menu_item_priority.should == 2
-      end
-    end
-    
-    describe "menu item display if" do
-      it "should be a proc always returning true if not set" do
-        config.menu_item_display_if.should be_instance_of(Proc)
-        config.menu_item_display_if.call.should == true
-      end
-      it "should be settable" do
-        config.menu :if => proc { false }
-        config.menu_item_display_if.call.should == false
-      end
-    end
-
     describe "route names" do
       it "should return the route prefix" do
         config.route_prefix.should == "admin"